requirements.txt
flask==2.0.1
flask-cors==4.0.0
werkzeug==2.0.3
yt-dlp==2023.11.16
requests==2.31.0
beautifulsoup4==4.12.2
flask-bootstrap==3.3.7.1
browser-cookie3==0.19.1
gunicorn==21.2.0
python-dotenv==1.0.0 